Move random-data utils to a separate file
Also, delete formatTime util as it's unused
